Move template state infrastructure into a dedicated states module
